+++ src/include/CompareOpt.h
@@ -6,14 +6,13 @@
 #define TEST_BISON_COMPAREOPT_H
 
 #include <string>
-#include "../include/sql/Query.h"
 #include "sql/Enums.h"
 
 using namespace std;
 
 class CompareOpt {
 public:
-    CompareOpt(CompareOperation opt);
+    CompareOpt(CompareOperation compareOpt);
 
     void setTheId(string theId);
     void setBool(bool setBoolVariable);
